Add cover position customization
Some photos have relevant content in different parts of the frame.
The previews can often strip these because it was defaulting to a
cover-center position. Here we let notes specify an optional dictionary
within the note metadata to customize the behavior of each featured
photo. Featured photos that are just string paths are still supported
and will default to center positioning.